requirements.txt
torchvision~=0.15.2
torch~=2.0.1
Pillow~=10.0.0